% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_non_chars.R
\name{check_non_chars}
\alias{check_non_chars}
\title{Checking columns for non-character variables in each row}
\usage{
check_non_chars(column)
}
\arguments{
\item{column}{A vector}
}
\value{
A sanitized vector
}
\description{
This function takes a long dataframe as input and returns the position of these variables
}
\examples{
check_non_chars(smpsn$companies)

}
