% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_col_names.R
\name{clean_col_names}
\alias{clean_col_names}
\title{Cleans column names in a dataframe}
\usage{
clean_col_names(column_names, keep_special)
}
\arguments{
\item{column_names}{A vector of column names}

\item{keep_special}{A boolean variable to keep/discard special characters}
}
\value{
A sanitized vector
}
\description{
This function takes a vector of characters and cleans them for abnormalities
}
\examples{
clean_col_names(column_names = colnames(sanir::smpsn), keep_special = FALSE)

}
