% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorCorrelation.R
\name{colorCorrelation}
\alias{colorCorrelation}
\title{Colors correlation matrix for GGally function}
\usage{
colorCorrelation(data, mapping, method = "pearson", use = "pairwise")
}
\arguments{
\item{data}{Dataframe being fed into ggpairs function, should be wide df}

\item{mapping}{mapping specifies columns in df being used}

\item{method}{method of correlation to be ran on df}

\item{use}{type of operation, pairwise will use rows for correlation}
}
\value{
a ggplot object
}
\description{
This function is used inside ggpairs in GGally package and uses a wide dataframe
}
\examples{
testCor <- tibble::tibble(ColA = seq(from = 1, to = 10), ColB = seq(from = -10, to = -1))
corrMatrix <- testCor \%>\%
GGally::ggpairs(axisLabels = "none", upper = list(continuous = colorCorrelation,
lower = list(continuous = GGally::wrap("points", size = 0.8))))

}
