% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_set.R
\name{cols_set}
\alias{cols_set}
\title{Checking columns for non-character variables in each row}
\usage{
cols_set(columns)
}
\arguments{
\item{columns}{A vector of column names}
}
\value{
A sanitized vector
}
\description{
This function takes a long dataframe as input and returns the position of these variables
}
\examples{
check_non_chars(smpsn$companies)

}
