% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLinePlot.R
\name{multiLinePlot}
\alias{multiLinePlot}
\title{Creates plotly linechart with multiple lines as colors}
\usage{
multiLinePlot(data, xAxis, yAxis, multi, titlePlot)
}
\arguments{
\item{data}{Dataframe being fed into plotly, should be long df}

\item{xAxis}{name of column to be used as x axis}

\item{yAxis}{name of column to be used as y axis}

\item{multi}{name of column that has seperate names for each line}

\item{titlePlot}{name of plot}
}
\value{
a plotly object
}
\description{
Plotly function with multiple lines
}
\examples{
sample <- tibble::tibble(time = seq(from = 1, to = 10), volume = seq(from = 11, to = 20),
car = "Toyota") \%>\% rbind(tibble::tibble(time = seq(from = 1, to = 10),
                                        volume = seq(from = 31, to = 40),
                                        car = "Jeep"))
multiLinePlot(data = sample, xAxis = "time", yAxis = "volume", multi = "car",
titlePlot = "Car Details")
}
