% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{returns}
\alias{returns}
\title{Checking columns for non-character variables in each row}
\usage{
returns(ticker, date)
}
\arguments{
\item{ticker}{A Yahoo Finance stock ticker}

\item{date}{Date where returns begin}
}
\value{
A plotly graph of returns
}
\description{
This function takes a stock ticker and finds daily returns from a specified date
}
\examples{
returns <- aapl

}
